# tsoa/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"LoginRequestParameter": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RegisterRequestParameter": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeAddLeaveRequestBody": {
				"properties": {
					"employee_id": {
						"type": "string"
					},
					"leave_type_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"duration": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"employee_id",
					"leave_type_id",
					"start_date",
					"duration"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeCreateRequestBody": {
				"properties": {
					"noInduk": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"birth_date": {
						"type": "string",
						"format": "date-time"
					},
					"phone_number": {
						"type": "string"
					},
					"employement_date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"noInduk",
					"name",
					"birth_date",
					"phone_number",
					"employement_date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmployeeUpdateRequestBody": {
				"properties": {
					"noInduk": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"birth_date": {
						"type": "string",
						"format": "date-time"
					},
					"phone_number": {
						"type": "string"
					},
					"employement_date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header",
				"flow": "implicit",
				"scopes": {
					"write:pets": "modify things",
					"read:pets": "read things"
				}
			}
		}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
		"/api/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequestParameter"
							}
						}
					}
				}
			}
		},
		"/api/auth/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Authorization"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RegisterRequestParameter"
							}
						}
					}
				}
			}
		},
		"/api/employee": {
			"get": {
				"operationId": "GetAllEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmployeeCreateRequestBody"
							}
						}
					}
				}
			}
		},
		"/api/employee/leave-stats/{id}": {
			"get": {
				"operationId": "GetEmployeLeaveStats",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/employee/add-leave": {
			"post": {
				"operationId": "AddLeave",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmployeeAddLeaveRequestBody"
							}
						}
					}
				}
			}
		},
		"/api/employee/{id}": {
			"put": {
				"operationId": "UpdateEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmployeeUpdateRequestBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEmployee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Employee"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/roles": {
			"get": {
				"operationId": "GetAllRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Roles"
				],
				"security": [
					{
						"api_key": [
							"read:role"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/tracker": {
			"get": {
				"operationId": "GetTracker",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Tracker"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "period",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/user-permission/get-all": {
			"get": {
				"operationId": "GetAllUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [
					{
						"api_key": [
							"read:user"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/user-permission/create": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [
					{
						"api_key": [
							"create:user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"roles": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"roles",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user-permission/update/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [
					{
						"api_key": [
							"update:user"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"roles": {
										"items": {
											"type": "string"
										},
										"type": "array"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"roles",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user-permission/delete/{id}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"User Permission"
				],
				"security": [
					{
						"api_key": [
							"delete:user"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"securitydefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "access_token",
			"in": "header",
			"flow": "implicit",
			"scopes": {
				"write:pets": "modify things",
				"read:pets": "read things"
			}
		}
	}
}